tdf#130857 qt weld: Add standard message dialog buttons defined in .ui
When processing a .ui file in QtBuilder, also evaluate the
GtkMessageDialog:buttons property [1] and add the corresponding
buttons to the QMessageBox.

This is needed e.g. for the dialog that can be triggered
as follows, for which support will be declared in an upcoming commit:

* start Math
* "Format" -> "Font Size"
* click the "Default" button in the dialog

(Without this commit, the dialog would only have a
single "OK" instead of a "Yes" and a "No" button.)

[1] https://docs.gtk.org/gtk3/property.MessageDialog.buttons.html

Change-Id: I69c9e736b5080a4a824ffeb5f364492402e406c8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181726
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>