tdf#130857 qt weld: Support "Printer Options" dialog
Declare support for the "Printer Options" dialog that
can be triggered as follows:

* "File" -> "Printer Settings"
* in the "Printer Setup" dialog that shows up, click the "Options"
  button

This means that native Qt widgets are used for that dialog
now when using the qt5 or qt6 VCL plugin and starting LO with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

For some modules, there are extra controls whose .ui files
are added in this commit as well.

Change-Id: I5f8fcb5207b5db8cd4c71a6698c215ae09d97dbc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181732
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>