tdf#130857 qt weld: Support "Delete {Header,Footer}?" dialogs
To trigger the "Delete Footer?" dialog

* start Writer
* "Insert" -> "Header and Footer" -> "Footer" -> "Default Page Style"
* type something in the footer
* "Insert" -> "Header and Footer" -> "Footer" -> "Default Page Style"
  to disable again

For "Delete Header?" one, choose "Header" instead of "Footer"
in the steps mentioned above.

This means that native Qt widgets are used for that dialog
now when using the qt5 or qt6 VCL plugin and starting LO with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

Change-Id: I38157fed85a05af74383d1d8f2b7548b323da876
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181737
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>