vcl: Drop LongCurrencyFormatter::SetDecimalDigits
It's only ever called with a value of 0 from the ctor.
Drop the method and LongCurrencyFormatter::mnDecimalDigits
and call ReformatAll() in the ctor directly (which was
previously called from LongCurrencyFormatter::SetDecimalDigits.

Let LongCurrencyFormatter::GetDecimalDigits simply return 0
for now. This could be simplified further to adjust all
callers.

Note that this class is a different class from
weld::LongCurrencyFormatter (defined in
include/vcl/weldutils.hxx).

Change-Id: I571f2edf79d1e203d5c05acf684a3e50e593f477
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181802
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>