cool#11064 Set priority of some idles to TaskPriority::DEFAULT_IDLE
LOK clients can set their "any input" callback to interrupt idle jobs
when they have something to execute which is less important that core
"high priority" jobs but is more important than "low priority" jobs.

For example calculating the layout for the visible area of a just
changed Writer document is high priority, while calculating the
non-visible area is low priority.

Experimenting with an upcoming change to only interrupt the scheduler
tasks when there are no high priority tasks scheduled, I noticed a
number of idles which have a default priority, which is rather high:
TaskPriority::DEFAULT is higher than a repaint priority.

Fix the problem by checking what idles prevent the early paint of the
visible area in Writer in LOK client code and by downgrading the
priority of these idles to TaskPriority::DEFAULT_IDLE. With this, the
"early paint of the visible Writer area" feature in LOK client code
continues to work, even if the any input callback starts to not
interrupt for the case when there are
TaskPriority::HIGHEST..TaskPriority::REPAINT tasks in the scheduler,
ready to be invoked.

Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181794
Reviewed-by: Michael Meeks <michael.meeks@collabora.com>
Change-Id: I42b3e0adbad98ef992ee3a454aba431a662a8126