weld: Let Formatter:m_aInputHdl use double instead of sal_Int64
Switch the param for the out param in Formatter:m_aInputHdl
from sal_Int64* to double* and adjust the methods/Links
that get set via Formatter::SetInputHdl.

Formatter itself otherwise already uses double for its
value (and min/max, etc.), so using double here as well
is consistent with that.

It also matches what SalInstanceSpinButton, one of the
classes that uses Formatter internally does since

    Change-Id: I31bc2b95b46bae8174388cddc9a383ed61e6763e
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon Feb 17 13:10:16 2025 +0100

        tdf#130857 Pass double values to/from weld::SpinButton subclasses

Formatter:m_aInputHdl gets used in Formatter::ImplGetValue
to convert text into a number. While Formatter:m_aInputHdl
was returning sal_Int64 so far, that was already converted
to a double value there anyways, depending on the number
of digits (s. call to `GetDecimalDigits()`), with basically
the same logic that the weld::SpinButton base class
does it by now since the above commit (while SalInstanceSpinButton
was implementing the logic previously.)

Notes on the methods/Links that are being passed via
Formatter::SetInputHdl:

* SpinfieldControl::ParseInputHdl already had a double value
  that it was multiplying by the same factor that
  Formatter::ImplGetValue was earlier then dividing by,
  so the multiplication there can simply be dropped.

* SalInstanceSpinButton now uses weld::convert_value_to_double
  to get a double value as needed.
  In an upcoming step, weld::SpinButton::parse_text will
  be adjusted to also use double right away, as another
  step to let its subclasses use double directly.
  (Preparing for this was the main motivation for this commit.)

* LongCurrencyFormatter can have decimals and now does the
  division previously done in Formatter::ImplGetValue
  to achieve the same result as previously.

* DateFormatter and TimeFormatter both use values that don't have
  any decimals, so the conversion should already
  have been a no-op (dividing by 1) and therefore doesn't
  need to be added elswhere.

In case anything was missed in a handler here, adding a
division by

    weld::SpinButton::Power10(m_xWidget->GetFormatter().GetDecimalDigits())

as was previously done in Formatter::ImplGetValue
should likely fix the issue.

Change-Id: Ibc297928f76ac1dab0080c36cef44ddddefab7b8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181804
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins