tdf#165089 sw: disable grid centering for DOC/DOCX compatibility
This change disables Writer CJK text grid centering when the
MS_WORD_COMP_GRID_METRICS compatibility flag is set.

Grid centering is useful for the Writer-native grid implementation,
which defines both line pitch and a line count, but does not make sense
for the Word grid algorithm which always fills a page at a specified
pitch.

Change-Id: Iaf63b2545636a678d5ba7264e0fb27e042e215c0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181829
Reviewed-by: Jonathan Clark <jonathan@libreoffice.org>
Tested-by: Jenkins