gtk: Port from deprecated gtk_widget_show
`gtk_widget_show` was deprecated in GTK 4.10 [1].
Replace with calls to `gtk_widget_set_visible` instead
(which itself just calls the deprecacted `gtk_widget_show`
internally for now in GTK 4 [2]) as recommended.

While it's not deprecated for GTK 3, also replace it in gtk3-only
code where the semantics are the same. Quoting from the
GTK 3 `gtk_widget_set_visible` doc [3]:

> This function simply calls gtk_widget_show() or gtk_widget_hide() but is
> nicer to use when the visibility of the widget depends on some
> condition.

A similar change will be done for `gtk_widget_hide`
in an upcoming commit.

[1] https://docs.gtk.org/gtk4/method.Widget.show.html
[2] https://gitlab.gnome.org/GNOME/gtk/-/blob/849b5e55d531a8a5645487853cb3876681674c31/gtk/gtkwidget.c#L5837-5859
[3] https://docs.gtk.org/gtk3/method.Widget.set_visible.html

Change-Id: Ib99a8c2420c50d918392070b10adb1aa5544492d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181833
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>