ScriptForge (SF_Calc) formatting methods
So far only the SetCellStyle() method
provided means to modify cell formats.

The actual patch introduces next methods
to apply simple and common formats. For
other options use the XCellRange
UNO interface.

- calc.AlignRange(targetrange,alignment,...)
- calc.BorderRange(targetrange,borders,...)
- calc.ColorizeRange(targetrange,foreground,background,...)
- calc.DecorateFont(targetrange,fontname,fontsize,
         decoration,...)
- calc.FormatRange(targetrange,numberformat,locale,...)

Extra arguments ("...") are:
  FilterFormula: a Calc formula to select cells
  FilterScope; CELL, ROW or COLUMN
to allow a kind of static conditional formats.

All methods are applicable to Basic and Python
user scripts.

This change will require an update of the
user documentation.

Change-Id: Ic824aebd763f2b55364c2c0bea17430f32512742
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181854
Tested-by: Jenkins
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>