Avoid short blinking of text when activating TextEdit
In Presentation with title/Body with text directly activating
one by clicking makes the other 'blink' briefly when it gets
deactivated. All that has to do with OverlayManager, repaints
and the text being 'hidden' (to not be shown twice during
edit). That needs to be synchronized since OM refreshes using
a timer and direct paint to the target, while repaint uses
system's invalidate-and-paint-callback functionality. It
has shown that the immediate flush of the OM that was used
before is not needed anymore, so I removed that.
Note that there is task tdf#102293 mentioned as reason for
the flush, so I checked that the there mentioned triple/
double-click on text does the right thing.

Change-Id: Id9539129c5d6989959974befd15300abdfef3186
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181904
Tested-by: Jenkins
Reviewed-by: Thorsten Behrens <thorsten.behrens@allotropia.de>