Remove OverlayManager::flush() mechanism
This is a follow-up to 'Avoid short blinking of text
when activating TextEdit' commit. The update using the
timer works nowadays flawless, so no longer needed.
I separated the two places in two commits because this
one handles some places I do not know good enough to
be completely sure, so it would be simpler to turn
back in case of (unexpected) problems.

Change-Id: I336515838949585b98101ff566e9137b898da82f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181906
Reviewed-by: Thorsten Behrens <thorsten.behrens@allotropia.de>
Tested-by: Jenkins