tdf#130857 qt weld: Format QtInstanceFormattedSpinButton value
Set a function used to convert the value to text to show in
the QtDoubleSpinBox used in QtInstanceFormattedSpinButton.

For further background, see also

    commit 6871b55b6915396caa3e5aca7e233f6a5efc7864
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Sat Feb 15 17:47:01 2025 +0100

        tdf#130857 qt weld: Implement SpinButton text <-> value conversion

and related commits doing similar for QtInstanceSpinButton.

This e.g. makes the time value shown in the Impress
"Slide Show" -> "Slide Show Settings" dialog be formatted
as expected in a WIP branch that declares support
for that dialog.

Conversion the other way around (i.e. text to value)
is still missing.

Change-Id: I229553c15ca318643472b9ca797f0a81060a3f2b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181914
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins