tdf#164487: Introduce "Show underline" MS Word compatibility option
It corresponds to ulTrailSpace element in OOXML. This commit creates
the option in officecfg, UI, ODF, and implements its initial support
in layout.

Unresolved problems (pre-existing, not created here):

1. Short lines show underline unconditionally.  In the unit test doc,
line 2 is such.  In SwTextGuess::Guess, there is the "first check if
everything fits to line" shortcut, that skips SwHolePortion creation,
that causes the problem.

2. The first character of the trailing whitespace is underlined.  It
is caused by AdjustCutPos only stepping back until there is a single
character left. But to fix that,  we need to handle the "zero-length
portion plus some extra" case.

Change-Id: Icc84fcff3ae62ced198522a80b8ce6b0e12e3a3c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182188
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182214
Tested-by: Mike Kaganski <mike.kaganski@collabora.com>