tdf#130857 qt weld: Support Svx(Decimal)NumberDialog
This means that native Qt widgets are used for that dialog
now when using the qt5 or qt6 VCL plugin and starting LO with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

The .ui file is used for 2 dialogs:

* SvxNumberDialog which handles integer values
  and uses "just" weld::SpinButton
* SvxDecimalNumberDialog which handles decimal numbers
  and uses weld::FormattedSpinButton

Sample steps to trigger the SvxDecimalNumberDialog:

* start Writer
* "Tools" -> "Options"
* "LibreOfficeDev" -> "Advanced" -> "Open Expert Configuration"
* Search for "ColScale" and double-click on the only search result
  (config entry "org.openoffice.Office.Calc" > "Filter" > "Import" >
   "MS_Excel", property "ColScale") to modify it

Handling for decimal numbers using QtInstanceFormattedSpinButton
still needs some tweaks that will be done in upcoming commits.
(For example, it currently rounds the value when double-clicking
on a proparty that has a floating point value set, and only allows
typing integer values, not actual floating point values.)

Change-Id: I25707ab6c5877eea2270bc21a27d08483d0f52e6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181941
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>