tdf#165322 sw: prefer insert/delete redline in mouse-over tooltip
Unfortunately SwCursorShell::GetContentAtPos() can only report 1 redline
for a given position; in case there is a format and a insert/delete
redline at the same position, the latter is more important.

This is then used in SwEditWin::RequestHelp() to display a tooltip if
the user moves the mouse cursor over a redline.

Change-Id: I32af7a154873e3fa5775a711980bb3dab3e307ad
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181960
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>