tdf#130857 qt weld: Implement spin button input validation
Override QDoubleSpinBox::validate [1] to validate
whether a given input text is valid.

If a handler for parsing the text is set and
it can successfully parse the text, return
QValidator::Acceptable, otherwise
QValidator::Intermediate (allowing to continue
editing/typing).

Without this, in the Impress "Slideshow" -> "Slideshow Settings"
dialog, modifying the value for the "Loop and repeat after"
spinbox wasn't possible by typing (as opposed to using the
up/down arrow keys or buttons) at all in a WIP branch declaring
support for that dialog, see upcoming

    Change-Id: I2d1552d4bc5a97da38ddb3678f20aa2a1faeb014
    Author: Michael Weghorn <m.weghorn@posteo.de>

        tdf#130857 qt weld: Support Impress slide show settings dialog

With this change in place, at least selecting a number
and typing another one to replace the existing one to
change the time works.
(Further improvements to come in another commit before
the dialog will actually be declared supported.)

[1] https://doc.qt.io/qt-6/qdoublespinbox.html#validate

Change-Id: If214119799fdf32f7295cacfab51d2c8ce5ba505
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181962
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins