tdf#130857 qt weld: Use QSpinBox in SpinButton's weld::Widget methods
In the implementation of the QtInstanceWidget methods,
don't use the `m_pWidget` member directly, but instead
call the virtual `getQWidget()` method to get the
widget to work with.

This ensures that the QSpinBox is used for QtInstanceSpinButton
and QtInstanceFormattedSpinButton, which pass the QSpinBox's
QLineEdit into the QtInstanceEntry base class ctor in
order to be able to implement the weld::Entry interface, see

    commit 12fb40fd7bc771813b825806cb4151833d7cd789
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Nov 27 23:40:16 2024 +0100

        tdf#130857 qt weld: Let SpinButton derive from Entry

Using the actual QSpinBox (i.e. the "main widget") e.g. makes
disabling the spinbox in the Impress "Slide Show" -> "Slide Show Settings"
dialog work as expected.

(Without this commit in place, it wasn't possible to type into
the "Loop and repeat after" SpinButton, but it was still possible
to change the value using the "up/down" buttons of the SpinButton
even if the checkbox was not checked.

Change-Id: I561a662830203b20db4d247fcc90416c13950e7e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181967
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins