tdf#130857: sd presentationsdialog.ui: Define widgets in order
Switch the order in which the widgets in the dialog's .ui file
are defined so that the order matches the visual appearance
(and "left-attach"/"top-attach" properties order), which makes sure that tab focus
order with the Qt-based VCL plugins is correct as well in that dialog when
using native Qt widgets (SAL_VCL_QT_USE_WELDED_WIDGETS=1), where focus order
is (by default) based on widget creation order.

See

    commit 02692566ad9fc7c3484f8581ffa0004cd4e43987
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Oct 24 17:43:35 2024 +0200

        tdf#130857 optnewdictionarydialog.ui: Define focusable widgets in order

for more background.

This is the Impress "Slide Show" -> "Slide Show Settings"
dialog.

Change-Id: I1324e5e5a410c49859eec14d6d4681a5a3339a65
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181969
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>