tdf#130857 qt weld: Add logic to find first treeview text column
The weld::TreeView::get_text doc says:

    // col index -1 gets the first text column

So far, the QtInstanceTreeView implementation was simply
assuming that  the first text column is the first column, i.e.
the column at index 0.

Add new method QtInstanceTreeView::firstTextColumnModelIndex
that tries to find the first text column in a given row
by finding the first column where data of type
Qt::DisplayRole ("The key data to be rendered in the form of text.
(QString)" [1]) is set.

Assert if no such column was found.

This at least works as expected for Calc's
"Sheet" -> "Show Sheet" dialog for which initial
support was added in

    commit 5d94ffde5125960fff0c177635292b2de5e20d38
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Nov 29 15:44:55 2024 +0100

        tdf#130857 qt weld: Support Calc's "Show Sheet" dialog

[1] https://doc.qt.io/qt-6/qt.html#ItemDataRole-enum

Change-Id: Ibba0f99d6be006d7b8d896b0bfc396cfabbec084
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181970
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>