tdf#130857 qt weld: Implement more treeview methods using first text col
Following

    Change-Id: Ibba0f99d6be006d7b8d896b0bfc396cfabbec084
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Feb 20 22:24:37 2025 +0100

        tdf#130857 qt weld: Add logic to find first treeview text column

, implement the remaining weld::TreeView methods whose
doc says that a column index of -1 refers to the first text
column. Also implement handling of other col indices where
that wasn't done yet.

Whether the logic introduced commit will turn out to be sufficient
in the end remains to be seen. (E.g. whether other text data are
always already set at the point in time that
QtInstanceTreeView::set_text gets called with a col index of -1,
which is a prerequisite for the logic to find the correct column.)

Change-Id: I49298245f2773309ceb33fe05971dbc1907c5675
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181971
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins