tdf#130857 qt weld: Let QStandardItemModel::itemFromIndex create item
QStandardItemModel::itemFromIndex "will lazily create an item for
the index (using itemPrototype()), and set it in the parent item's
child table, if no item already exists at that index." [1].

Make use of that instead of manually calling the QStandardItem
ctor.

The main motivation is that this simplifies an upcoming change.

[1] https://doc.qt.io/qt-6/qstandarditemmodel.html#itemFromIndex

Change-Id: I371cdf8bd46b5f640c57760e4b52dda6489876c6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182005
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>