tdf#130857 qt weld: Add support for special toggle button column
As the comment for weld::TreeView in weld.hxx says:

    /* Model column indexes are considered to begin at 0, but with special columns
       before index 0. A expander image column (and an additional optional toggle
       button column when enable_toggle_buttons is used). Column index -1 is
       reserved to access those columns.
    */
    class VCL_DLLPUBLIC TreeView : virtual public Widget

Add initial support for the special toggle button
column (only for checkbuttons/checkboxes initially,
not radiobuttons).

That special column gets enabled when
weld::TreeView::enable_toggle_buttons gets called.
For those methods that provide access, a column
index of -1 is used to access that special column,
and the first "normal" column uses an (external)
column index of 0.

In the QtInstanceTreeView's QStandardItemModel however,
the special column has an index of 0 (because it
is the first one) and the next column has a column
index of 1.

Adjust QtInstanceTreeView::modelIndex to calculate
the model index accordingly when the special
column has been enabled.

This is similar to what the GTK and VCL implementations
do, see GtkInstanceTreeView::to_internal_model
or SalInstanceTreeView::to_internal_model.

Introduce QtInstanceTreeView::m_bExtraToggleButtonColumnEnabled
to keep track of whether this column has been enabled.

Add QtInstanceTreeView::toggleButtonModelIndex
for a helper to get the QModelIndex for the cell
containing the checkbox for a given row.
(For now, this assumes that the column index is 0,
which may or may not change when adding support for
the other "special column" for the expander image
also mentioned in the weld::TreeView doc quoted above.

In QtInstanceTreeView::enable_toggle_buttons, resize
the column according to the contents, as otherwise
it was taking way too much space when tested with
an upcoming commit adding support for Writer's
"Tools" -> "Spelling" -> "Options" dialog.

Change-Id: I5889429bf77e69cec4004fd54b2f93454d911c36
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182006
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>