remove unused LoadColorsFromTheme() from gtk theming code
the usual theming approach is to set the colors at the source i.e.
using the widget toolkit's mechanism and then the UpdateSettings
function should be able to extract that color and set it in the
StyleSettings.

LoadColorsFromTheme like functions are used where the widget toolkit
doesn't provide a nice way of setting the colors at source.

Change-Id: I235f4dc06d9f8607cad6bfcdafae66f597f10547
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182003
Reviewed-by: Sahil Gautam <sahil.gautam.extern@allotropia.de>
Tested-by: Jenkins