tdf#130857 qt weld: Support spelling options dialog
Declare support for the .ui files used for Writer's
"Tools" -> "Spelling" -> "Options" dialog.

This means that native Qt widgets are used for that dialog
now when using the qt5 or qt6 VCL plugin and starting LO with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

This dialog makes use of the weld::TreeView feature
of having a special checkbutton column for which support
was added in the QtInstanceTreeView implementation
in previous commit

    Change-Id: I5889429bf77e69cec4004fd54b2f93454d911c36
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Feb 21 15:20:49 2025 +0100

        tdf#130857 qt weld: Add support for special toggle button column

Change-Id: Ic0b81e9e9d429ad558783121f6981719ee8fe920
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182007
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>