tdf#130857 qt weld: Don't assert text has been set when getting it
If QStandardItemModel::data returns a QVariant that cannot
be converted to QString (most likely an invalid QVariant
because no data has explicitly been set), simply return an empty
string QtInstanceTreeView::get_text and don't assert.

This assert was e.g. triggered for the "Tools" -> "Spelling"
-> "Options" -> "Edit" dialog in Writer in a WIP branch
where support for that dialog is declared.

Change-Id: Ic5782c418758becdc636356c13ab878f5fd2f007
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182012
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins