tdf#130857 weld: Make {Icon,Tree}View::{,un}select_all purely virtual
No longer let the base class implementation
`weld::TreeView::select_all()` call `weld::Treeview::unselect(-1)`
and similar, but make them purely virtual and implement in
the subclasses instead.

This introduces a working implementation for QtInstanceTreeView,
where the special value of -1 wasn't handled in the `select()` and
`unselect()` implementations.
(It also doesn't feel very "natural"/obvious to me and wasn't explicitly
documented for `select` and `unselect`.)

For the GTK and VCL implementation, leave the logic as it was
for maximum backward compatibility for now.

I noticed this in a WIP branch declaring support for
the "Tools" -> "Spelling" -> "Options" -> "Edit" dialog in Writer
where typing a new word not yet in the list would not result in
the previously selected word in the treeview to become unselected
despite SvxEditDictionaryDialog::ModifyHdl calling

    m_pWordsLB->unselect_all();

Change-Id: I2f681e2297a553beedd10fd7054adc539ac13ddc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182013
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>