tdf#130857 qt weld: Also select item in QtInstanceTreeView::set_cursor
Align the Qt implementation with `gtk_tree_view_set_cursor`
(used in GtkInstanceTreeView::set_cursor), which [1]

> Sets the current keyboard focus to be at path, and selects it.

So far, QtInstanceTreeView::set_cursor was only setting the
cursor (i.e. the item that has keyboard focus once focus moves
to the tree view), but not updating the selection.

The different behavior was seen in a WIP branch declaring support
for the "Tools" -> "Spelling" -> "Options" -> "Edit" dialog in Writer
where typing a word already in the list would not result in
that word to become selected with qt6 and SAL_VCL_QT_USE_WELDED_WIDGETS=1
while it would with gen or gtk3.

[1] https://docs.gtk.org/gtk3/method.TreeView.set_cursor.html

Change-Id: I1b916f7f176e0a4862f8e6bee720a694e50b2972
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182017
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins