tdf#130857 qt weld: Convert Qt to VCL accelerator in get_label()
In QtInstanceLabel::get_label and QtInstanceButton::get_label,
convert the text potentially containing a Qt accelerator "&"
and escaped "&" ("&&") to the VCL variant using "~" as accelerator.
(The same is already done the other way around in `set_label()`).

Without this, the "New" button in the dialog triggered as follows
in a WIP branch where support for that dialog is declared for
SAL_VCL_QT_USE_WELDED_WIDGETS=1 was displayed with a literal "&New"
as its text, because SvxEditDictionaryDialog::SvxEditDictionaryDialog
does this:

    sNew = m_xNewReplacePB->get_label();
    // ...
    m_xNewReplacePB->set_label(sNew);

To trigger the dialog:

* "Tools" -> "Spelling" -> "Options"
* select the Hungarian dictionary
* click the "Edit" button

Change-Id: Ia457f0d8c3dc03b33fa3cb026814830b18c5ca8a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182021
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins