tdf#130857 qt: Don't expect XFilePicker3 template in XFolderPicker2 init
QtFilePicker implements both, the XFilePicker3 and the XFolderPicker2
interfaces.

While a sal_Int16 specifying the TemplateDescription to use
is passed to the initialize method in case of XFilePicker3
(s. FileDialogHelper_Impl::FileDialogHelper_Impl), that is
not the case for XFolderPicker2 (s. sfx2::createFolderPicker,
frame 2 in the below sample backtrace; a sal_Int32 with value 0
is passed there instead).

However, QtFilePicker was always expecting it and throwing an
IllegalArgumentException and thus not executing the rest of the
initialize method (but the folder picker would still show up and
work otherwise).

This can for example be seen for the following scenario:

* start Writer
* "Tools" -> "Options" -> "Security" -> "Macro Security"
* in the "Macro Security" dialog, switch to the "Trusted Sources"
  tab
* Click on the "Add" button in the "Trusted File Locations" section
  to open a folder picker.

This is a preexisting issue. But in an upcoming commit that
implements QtInstanceWindow::GetXWindow and the use of native
Qt widgets for that macro security dialog, the fact that the parent
widget isn't set (which would only happen further down in
QtFilePicker::initialize) would now have the practical implication
that the file dialog isn't properly shown as modal and blocking
the macro security dialog that is its parent.

Fix this by no longer throwing an exception when the
first arg is not a template ID, but simply make the
code using it conditional on that and skip it otherwise.

Also extract it to a new helper method
QtFilePicker::applyTemplate for readability.

(For comparison, note how for gtk3 the XFilePicker3 implementation
in SalGtkFilePicker is separate from the XFolderPicker2 implementation
in SalGtkFolderPicker and only the former tries to extract a
template ID.)

Sample backtrace:

    1 QtFilePicker::initialize QtFilePicker.cxx 730 0x7fc196b45652
    2 sfx2::createFolderPicker filedlghelper.cxx 1217 0x7fc1a6544781
    3 MacroSecurityTrustedSourcesTP::AddLocPBHdl macrosecurity.cxx 303 0x7fc158926afc
    4 MacroSecurityTrustedSourcesTP::LinkStubAddLocPBHdl macrosecurity.cxx 298 0x7fc158926a2d
    5 Link<weld::Button&, void>::Call link.hxx 105 0x7fc1a1426331
    6 weld::Button::signal_clicked weld.hxx 1558 0x7fc1a1401fdc
    7 SalInstanceButton::ClickHdl salvtables.cxx 2955 0x7fc1a13d32b8
    8 SalInstanceButton::LinkStubClickHdl salvtables.cxx 2943 0x7fc1a13d280d
    9 Link<Button *, void>::Call link.hxx 105 0x7fc1a0cd5881
    10 Button::Click()::$_0::operator()() const button.cxx 130 0x7fc1a0ccd082
    11 std::__invoke_impl<void, Button::Click()::$_0&>(std::__invoke_other, Button::Click()::$_0&) invoke.h 61 0x7fc1a0ccd055
    12 std::__invoke_r<void, Button::Click()::$_0&>(Button::Click()::$_0&) invoke.h 111 0x7fc1a0ccd005
    13 std::_Function_handler<void(), Button::Click()::$_0>::_M_invoke std_function.h 290 0x7fc1a0cccf2d
    14 std::function<void()>::operator() std_function.h 591 0x7fc1a0cf700e
    15 Control::ImplCallEventListenersAndHandler ctrl.cxx 307 0x7fc1a0cf51cb
    16 Button::Click button.cxx 130 0x7fc1a0cb98d8
    17 PushButton::Tracking button.cxx 1321 0x7fc1a0cbf71b
    18 vcl::Window::EndTracking window2.cxx 341 0x7fc1a0c6a763
    19 ImplHandleMouseEvent winproc.cxx 713 0x7fc1a0ca4aae
    20 ImplHandleSalMouseButtonUp winproc.cxx 2351 0x7fc1a0ca9041
    21 ImplWindowFrameProc winproc.cxx 2700 0x7fc1a0ca7d55
    22 SalFrame::CallCallback salframe.hxx 311 0x7fc196b6ebac
    23 QtFrame::CallCallback QtFrame.hxx 235 0x7fc196b6bc05
    24 QtWidget::handleMouseButtonEvent QtWidget.cxx 117 0x7fc196c7c5ca
    [...]

Change-Id: I008cb851d1ec1f9383e0f1e4342a2bb2a05e5c41
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182023
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>