tdf#130857 qt weld: Introduce QtXWindow
Similar to how gtk3's SalGtkXWindow, introduce
QtXWindow as a class implementing XWindow via
the TransportAsXWindow subclass.

Implement QtInstanceWindow::GetXWindow
to return an instance of this new class.

In QtFilePicker::initialize, extract the
QWidget parent from the QtXWindow if the
parent is one.

This is used e.g. in a WIP branch adding
support for the "Macro Security" dialog
using native Qt widgets
(SAL_VCL_QT_USE_WELDED_WIDGETS=1) in the
following scenario to set the proper
parent for the file dialog.

* start Writer
* "Tools" -> "Options" -> "Security" -> "Macro Security"
* in the "Macro Security" dialog, switch to the "Trusted Sources"
  tab
* Click on the "Add" button in the "Trusted File Locations" section
  to open a folder picker.

Change-Id: I55637876c9b93bbbfb63459caba6c89dfd4188a9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182024
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>