tdf#130857 Define widgets in XML Filter dlg .ui files in order
Switch the order in which the widgets in the .ui files
are defined so that the order matches the visual appearance
(and "left-attach"/"top-attach" properties order), which makes sure
that tab focus order with the Qt-based VCL plugins is correct as well
when using native Qt widgets (SAL_VCL_QT_USE_WELDED_WIDGETS=1), where
focus order is (by default) based on widget creation order.

See

    commit 02692566ad9fc7c3484f8581ffa0004cd4e43987
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Oct 24 17:43:35 2024 +0200

        tdf#130857 optnewdictionarydialog.ui: Define focusable widgets in order

for more background.

The dialog can be triggered in Writer via
"Tools" -> "XML Filter Settings" -> "New".

Change-Id: Icf1043da9d97efed6f1a5e1d3cb722d66f86921c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182031
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins