Make "XML Filter Settings" dialog modal and set parent
So far, the "Tools" -> "XML Filter Settings" dialog
did not have a parent and was not modal.

As a consequence (tested on Plasma Wayland):

* With qt6 or gen, the dialog would just "disappear" behind
  the main Writer window when clicking on the main Writer window
  or moving focus there using Alt+Tab.
* With gtk3, the dialog would always stay on top of the Writer
  window, but would still allow to interact with the Writer window,
  so when typing text, the cursor could easily move to be behind
  the dialog.

Make the dialog modal and set a parent (using
XMLFilterDialogComponent::initialize) , which makes the dialog
consistently appear on top of the Writer main window and the
Writer window insensitive/disabled until the dialog gets closed again.

Change-Id: I79ebc20cc3c059f217de9a1a909a7c7a424ab5ca
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182034
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>