lok: handle saveas as export
SaveAs saves the document to a new path.
In the jail, this is unhelpful, since
we upload the document after saving
from a specific path.

Instead, we do SaveACopy, or export,
which saves the given document as a
copy, retaining the original path
for subsequent saves.

This fixes a serious issue where
after hitting Ctrl+Shift+S, the
document would no longer store any
further changes (in fact, it was
saving the changes to a different
path, which was not used for
uploading).

Signed-off-by: Ashod Nakashian <ashod.nakashian@collabora.co.uk>
Change-Id: I6ba72df6a0a908d181786c02cb7c060177cd0302
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182071
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>
(cherry picked from commit fcf4dd4f09018874ad25a035815f46705db0740a)