tdf#130857 qt weld: Introduce QtInstanceAssistant
Add new class QtInstanceAssistant as the weld::Assistant
implementation using native Qt widgets.

A QWizard is used for the widget.

This commit introduces QtInstanceAssistant and adds an
initial implementation for most of its methods.
(Others still need to be added in the future and
calling them currently triggers an assert.)

While GtkAssistant [2] and weld::Assistant use
indices to identify pages, QWizard uses IDs. For now,
assume that those are equivalent. Some extra logic may
be needed to ensure this in the future (or an implementation
that is not built on this assumption).
The OUString identifier used in the weld::Assistant
API is something different and set as the QObject::objectName
property of the pages.

In QtBuilder, create a QWizard widget when encountering
a "GtkAssistant" object in a .ui file.

[1] https://doc.qt.io/qt-6/qwizard.html
[2] https://docs.gtk.org/gtk3/class.Assistant.html

Change-Id: I91b44da40736f6e61c64167d00305a1d7ba06b45
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182036
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>